pyproject.toml
[tool.poetry]
name = "devman-bot"
version = "0.1.0"
description = ""
authors = ["Andrey Miskov <amiskov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
environs = "^9.5.0"
backoff = "^2.2.1"
python-telegram-bot = "^20.6"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
